solve bug, few corrections
