+++ app/components/about/aboutme.module.css
@@ -35,7 +35,6 @@
   font-style: italic;
   font-size: 14px;
   margin-bottom: 50px;
-  background: antiquewhite;
 }
 
 .designerIntroduction {
