+++ app/pages/projetos/index.js
@@ -1,13 +1,11 @@
 import layoutStyles from "../../components/layout/layout.module.css";
-import ProjectMenu from "../../components/projects/project_menu/ProjectMenu";
-import ProjectsG from "../../components/projects/gallery/ProjectsG";
+import ProjectMenu from "../../components/projects/ProjectMenu";
 
 export default function Projects() {
   return (
     <div className={layoutStyles.informationSection}>
       <h1 className={layoutStyles.subtitles}>PROJECTOS</h1>
       <ProjectMenu />
-      <ProjectsG />
     </div>
   );
 }
